web.xml有註冊濾器Article不要砍，感謝 Article、ArticleSub、ArticleCategory的HDAO改用濾器，新增點擊數功能完成，群聊功能處理中

+++ src/main/java/com/cooklab/member_order/model/MemberOrderVO.java
@@ -31,7 +31,7 @@ public class MemberOrderVO implements Serializable{
 	@JoinColumn(name = "member_id", referencedColumnName = "member_id", insertable = false, updatable = false)
 	private MembersVO  members;
 //	=============WCC================	
-	@Column(name = "order_status", columnDefinition = "tinyint")
+	@Column(name = "order_status", columnDefinition = "int")
 	private byte orderStatus;  
 //	==============WCC=============================
 	@Column(name = "total_order_amount")
